+++ commons/ava_testsuite/staking_gecko_network_tests.go
@@ -12,15 +12,21 @@ import (
 	"github.com/kurtosis-tech/kurtosis/commons/testsuite"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
+	"strconv"
 	"time"
 )
 
 const (
-	USERNAME = "test"
-	PASSWORD = "test34test!23"
-	SEED_AMOUNT = 50000000000000
-	STAKE_AMOUNT = 20000000000000
-	NODE_SERVICE_ID = 0
+	STAKER_USERNAME           = "staker"
+	STAKER_PASSWORD           = "test34test!23"
+	DELEGATOR_USERNAME           = "delegator"
+	DELEGATOR_PASSWORD           = "test34test!23"
+	SEED_AMOUNT               = int64(50000000000000)
+	STAKE_AMOUNT              = int64(30000000000000)
+	DELEGATOR_AMOUNT              = int64(30000000000000)
+	NODE_SERVICE_ID           = 0
+	DELEGATOR_NODE_SERVICE_ID = 1
+
 	NORMAL_NODE_CONFIG_ID = 0
 
 	// The configuration ID of a service where all servies made with this configuration will have the same cert
@@ -31,32 +37,52 @@ const (
 type StakingNetworkRpcWorkflowTest struct{}
 func (test StakingNetworkRpcWorkflowTest) Run(network interface{}, context testsuite.TestContext) {
 	castedNetwork := network.(ava_networks.TestGeckoNetwork)
-	referenceNodeClient, err := castedNetwork.GetGeckoClient(NODE_SERVICE_ID)
+	stakerClient, err := castedNetwork.GetGeckoClient(NODE_SERVICE_ID)
 	if err != nil {
-		context.Fatal(stacktrace.Propagate(err, "Could not get reference client"))
+		context.Fatal(stacktrace.Propagate(err, "Could not get staker client"))
 	}
-	nodeId, err := referenceNodeClient.AdminApi().GetNodeId()
+	delegatorClient, err := castedNetwork.GetGeckoClient(DELEGATOR_NODE_SERVICE_ID)
 	if err != nil {
-		context.Fatal(stacktrace.Propagate(err, "Could not get reference nodeId."))
+		context.Fatal(stacktrace.Propagate(err, "Could not get delegator client"))
 	}
-	highLevelGeckoClient := ava_networks.NewHighLevelGeckoClient(
-		referenceNodeClient,
-		USERNAME,
-		PASSWORD)
-	_, err = highLevelGeckoClient.CreateAndSeedXChainAccountFromGenesis(SEED_AMOUNT)
+	stakerNodeId, err := stakerClient.AdminApi().GetNodeId()
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not get staker node ID."))
+	}
+	delegatorNodeId, err := stakerClient.AdminApi().GetNodeId()
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not get delegator node ID."))
+	}
+	highLevelStakerClient := ava_networks.NewHighLevelGeckoClient(
+		stakerClient,
+		STAKER_USERNAME,
+		STAKER_PASSWORD)
+	highLevelDelegatorClient := ava_networks.NewHighLevelGeckoClient(
+		delegatorClient,
+		DELEGATOR_USERNAME,
+		DELEGATOR_PASSWORD)
+	stakerXchainAddress, err := highLevelStakerClient.CreateAndSeedXChainAccountFromGenesis(SEED_AMOUNT)
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not seed XChain account from Genesis."))
 	}
-	pchainAddress, err := highLevelGeckoClient.TransferAvaXChainToPChain(SEED_AMOUNT)
+	stakerPchainAddress, err := highLevelStakerClient.TransferAvaXChainToPChain(SEED_AMOUNT)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not transfer AVA from XChain to PChain account information"))
+	}
+	_, err = highLevelDelegatorClient.CreateAndSeedXChainAccountFromGenesis(SEED_AMOUNT)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not seed XChain account from Genesis."))
+	}
+	delegatorPchainAddress, err := highLevelDelegatorClient.TransferAvaXChainToPChain(SEED_AMOUNT)
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not transfer AVA from XChain to PChain account information"))
 	}
 	// Adding stakers
-	err = highLevelGeckoClient.AddValidatorOnSubnet(nodeId, pchainAddress, STAKE_AMOUNT)
+	err = highLevelStakerClient.AddValidatorOnSubnet(stakerNodeId, stakerPchainAddress, STAKE_AMOUNT)
 	if err != nil {
-		context.Fatal(stacktrace.Propagate(err, "Could not add staker %s to default subnet.", nodeId))
+		context.Fatal(stacktrace.Propagate(err, "Could not add staker %s to default subnet.", stakerNodeId))
 	}
-	currentStakers, err := referenceNodeClient.PChainApi().GetCurrentValidators(nil)
+	currentStakers, err := stakerClient.PChainApi().GetCurrentValidators(nil)
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not get current stakers."))
 	}
@@ -64,8 +90,28 @@ func (test StakingNetworkRpcWorkflowTest) Run(network interface{}, context tests
 	actualNumStakers := len(currentStakers)
 	expectedNumStakers := 6
 	context.AssertTrue(actualNumStakers == expectedNumStakers, stacktrace.NewError("Actual number of stakers, %v, != expected number of stakers, %v", actualNumStakers, expectedNumStakers))
-	// TODO TODO TODO Test adding delegators
-	// TODO TODO TODO Test transferring staking rewards back to XChain
+	// Adding delegators
+	err = highLevelDelegatorClient.AddDelegatorOnSubnet(stakerNodeId, delegatorPchainAddress, DELEGATOR_AMOUNT)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Could not add delegator %s to default subnet.", delegatorNodeId))
+	}
+	/*
+		Currently no way to verify rewards for stakers and delegators because rewards are
+		only paid out at the end of the staking period, and the staking period must last at least
+		24 hours. This is far too long to be able to test in a CI scenario.
+	 */
+	remainingStakerAva := SEED_AMOUNT - STAKE_AMOUNT
+	_, err = highLevelStakerClient.TransferAvaPChainToXChain(stakerPchainAddress, stakerXchainAddress, remainingStakerAva)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Failed to transfer Ava from PChain to XChain."))
+	}
+	xchainAccountInfo, err := stakerClient.XChainApi().GetBalance(stakerXchainAddress, ava_networks.AVA_ASSET_ID)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Failed to get account info for account %v.", stakerXchainAddress))
+	}
+	actualRemainingAva := xchainAccountInfo.Balance
+	expectedRemainingAva := strconv.FormatInt(remainingStakerAva, 10)
+	context.AssertTrue(actualRemainingAva == expectedRemainingAva, stacktrace.NewError("Actual remaining Ava, %v, != expected remaining Ava", actualRemainingAva, expectedRemainingAva))
 }
 func (test StakingNetworkRpcWorkflowTest) GetNetworkLoader() (testsuite.TestNetworkLoader, error) {
 	return getFiveNodeStakingLoader()
@@ -83,6 +129,19 @@ func (test FiveNodeStakingNetworkFullyConnectedTest) Run(network interface{}, co
 	allServiceIds := castedNetwork.GetAllBootServiceIds()
 	allServiceIds[NODE_SERVICE_ID] = true
 
+	extraStakerClient, err := castedNetwork.GetGeckoClient(NODE_SERVICE_ID)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Failed to get extra staker client."))
+	}
+	highLevelExtraStakerClient := ava_networks.NewHighLevelGeckoClient(
+		extraStakerClient,
+		STAKER_USERNAME,
+		STAKER_PASSWORD)
+	err = highLevelExtraStakerClient.GetFundsAndStartValidating(SEED_AMOUNT, STAKE_AMOUNT)
+	if err != nil {
+		context.Fatal(stacktrace.Propagate(err, "Failed to add extra staker."))
+	}
+
 	// collect set of IDs in network
 	nodeIdSet := map[string]bool{}
 	for serviceId, _ := range allServiceIds {
@@ -99,33 +158,29 @@ func (test FiveNodeStakingNetworkFullyConnectedTest) Run(network interface{}, co
 
 	logrus.Debugf("Network ID Set: %+v", nodeIdSet)
 
-	// verify peer lists have set of IDs in network, except their own
+	// verify bootstrapper peer lists have full set of IDs in network.
 	for serviceId, _ := range allServiceIds {
 		client, err := castedNetwork.GetGeckoClient(serviceId)
 		if err != nil {
 			context.Fatal(stacktrace.Propagate(err, "Could not get client for service with ID %v", serviceId))
 		}
 		peers, err := client.AdminApi().GetPeers()
+		nodeId, err := client.AdminApi().GetNodeId()
 		if err != nil {
 			context.Fatal(stacktrace.Propagate(err, "Could not get peers of service with ID %v", serviceId))
 		}
-		logrus.Debugf("Peer set: %+v", peers)
+		logrus.Debugf("Length of Peer set for node %v: %v", serviceId, len(peers))
 
 		peerSet := map[string]bool{}
 		for _, peer := range peers {
 			peerSet[peer.Id] = true
-			// verify that peer is inside the nodeIdSet
-			context.AssertTrue(nodeIdSet[peer.Id], stacktrace.NewError("Peer ID %v is not in the node ID set", peer.Id))
 		}
-		// verify that every other peer (besides the node itself) is represented in the peer list.
-		actualNumPeers := len(peerSet)
-		expectedNumPeers := len(allServiceIds) - 1
-		context.AssertTrue(
-			actualNumPeers == expectedNumPeers,
-			stacktrace.NewError(
-				"Actual length of peer set, %v, is not equal to expected length of peer set, %v",
-				actualNumPeers,
-				expectedNumPeers))
+		for expectedNodeId, _ := range nodeIdSet {
+			// Nodes are expected to have all other nodes in the network on their peer list.
+			if nodeId != expectedNodeId {
+				context.AssertTrue(peerSet[expectedNodeId], stacktrace.NewError("Didn't find node ID %v in peer set", expectedNodeId))
+			}
+		}
 	}
 }
 
@@ -134,7 +189,7 @@ func (test FiveNodeStakingNetworkFullyConnectedTest) GetNetworkLoader() (testsui
 }
 
 func (test FiveNodeStakingNetworkFullyConnectedTest) GetTimeout() time.Duration {
-	return 30 * time.Second
+	return 60 * time.Second
 }
 
 // =============== Get Validators Test ==================================
@@ -385,8 +440,10 @@ func getFiveNodeStakingLoader() (testsuite.TestNetworkLoader, error) {
 		ava_services.LOG_LEVEL_DEBUG,
 		true,
 		serviceConfigs,
+		// TODO TODO Make this configurable by each test so that every test doesn't get a delegator!!
 		map[int]int{
-			NODE_SERVICE_ID: NORMAL_NODE_CONFIG_ID,
+			NODE_SERVICE_ID:           NORMAL_NODE_CONFIG_ID,
+			DELEGATOR_NODE_SERVICE_ID: NORMAL_NODE_CONFIG_ID,
 		},
 		2,
 		2)