+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 * Implement test for transferring assets between XChain accounts
 * Implement test for transferring assets from XChain to PChain
 * Remove `FiveNodeStakingNetworkBasicTest` (wasn't being used)
-* Expanding RpcWorkflow test to add a staker
+* Fixed fully connected test and added nonbootstrap node as staker
 
 ### Duplicate Node ID Test
 * Created `GeckoCertProvider` interface that's fed into the `GeckoServiceInitializerCore`, allowing for test writers to customize the certs that the certs get
@@ -23,6 +23,11 @@
 * Removed `FixedGeckoNetwork` in favor of `TestGeckoNetwork`, which allows for more control over the testnet that gets created
 * Removed the single-node and ten-node Gecko tests; they don't actually test anything useful when compared to the staking network tests
 * Test if the network functions as expected when nodes with duplicate node IDs occur
+
+### Full RPC Workflow Test
+* Expanded RpcWorkflow test to add a staker
+* Expanded RpcWorkflow test to add a delegator and transfer funds back to XChain
+* Created high level function to both fund and add a staker to default subnet
 
 # 0.2.1
 * Fix tee suppressing exit code of the Docker image